LUCENE-2386: IndexWriter commits unnecessarily on fresh Directory (take #2)
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@933613 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/test/org/apache/lucene/store/TestWindowsMMap.java
@@ -66,17 +66,22 @@ public class TestWindowsMMap extends LuceneTestCase {
     new File(TEMP_DIR,"testLuceneMmap").getAbsolutePath();
 
   public void testMmapIndex() throws Exception {
-    FSDirectory storeDirectory;
-    storeDirectory = new MMapDirectory(new File(storePathname), null);
-
+    // sometimes the directory is not cleaned by rmDir, because on Windows it
+    // may take some time until the files are finally dereferenced. So clean the
+    // directory up front, or otherwise new IndexWriter will fail.
+    File dirPath = new File(storePathname);
+    rmDir(dirPath);
+    MMapDirectory dir = new MMapDirectory(dirPath, null);
+    
     // plan to add a set of useful stopwords, consider changing some of the
     // interior filters.
     StandardAnalyzer analyzer = new StandardAnalyzer(TEST_VERSION_CURRENT, Collections.emptySet());
     // TODO: something about lock timeouts and leftover locks.
-    IndexWriter writer = new IndexWriter(storeDirectory, new IndexWriterConfig(
+    IndexWriter writer = new IndexWriter(dir, new IndexWriterConfig(
         TEST_VERSION_CURRENT, analyzer)
         .setOpenMode(OpenMode.CREATE));
-    IndexSearcher searcher = new IndexSearcher(storeDirectory, true);
+    writer.commit();
+    IndexSearcher searcher = new IndexSearcher(dir, true);
     
     for(int dx = 0; dx < 1000; dx ++) {
       String f = randomField();
@@ -87,14 +92,16 @@ public class TestWindowsMMap extends LuceneTestCase {
     
     searcher.close();
     writer.close();
-                rmDir(new File(storePathname));
+    rmDir(dirPath);
   }
 
-        private void rmDir(File dir) {
-          File[] files = dir.listFiles();
-          for (int i = 0; i < files.length; i++) {
-            files[i].delete();
-          }
-          dir.delete();
-        }
+  private void rmDir(File dir) {
+    if (!dir.exists()) {
+      return;
+    }
+    for (File file : dir.listFiles()) {
+      file.delete();
+    }
+    dir.delete();
+  }
 }
